ebpf: run TestProgramTestRunInterrupt only on Linux
TestProgramTestRunInterrupt tests Linux specific behaviour around signal
delivery. It doesn't make sense to run it on a different platform.

The test was copied without making any changes.
